Agrego la recuperación de estadísticas desde un archivo TOML

+++ server/game/stats/mana.h
@@ -6,18 +6,18 @@
 
 class Mana: public RecoverableStat {
 private:
-    int intelligence;
-    int factor_class_meditation;
+    uint8_t intelligence;
+    float factor_class_meditation;
 
 public:
-    Mana(int recovery_factor, int factor_class, int factor_race, int factor_class_meditation,
-         int intelligence, uint8_t level);
+    Mana(uint8_t recovery_factor, float factor_class, float factor_race, float factor_class_meditation,
+         uint8_t intelligence, uint8_t level);
 
-    void update_max(int level) override;
+    void update_max(uint8_t level) override;
 
     void meditate();
 
-    bool loose(int amount) override;
+    bool loose(uint16_t amount) override;
 };
